// frontend/src/app/library/components/lib-field.component/common/types.enum.ts
export enum FieldTypes {
    Button = 'button',
    Text = 'text',
    Date = 'date',
    ZIP = 'zip',
    Year = 'year',
    Money = 'money',
    Comment = 'comment',
    Selector = 'selector',
    Integer = 'integer',
    Check = 'check',
    Dropdown = 'dropdown',
    Autocomplete = 'autocomplete',
    Multicomplete = 'multicomplete',
    Password = 'password',
    Radio = 'radio',
    File = 'file',
    Phone = 'phone',
    Fax = 'fax',
    Number = 'number'
}